refactor(raijin): colocate runtime type contracts
