+++ yarn/raijin/src/initializer/interfaces.ts
@@ -1,4 +1,4 @@
-import type { FetchLike }         from '../runtime/download.interfaces.js'
+import type { FetchLike }         from '../runtime/download.js'
 import type { YarnCommandRunner } from '../yarn/interfaces.js'
 
 export interface RunRaijinInitializerOptions {
