--- yarn/raijin/src/runtime/download.interfaces.ts
@@ -1,1 +1,0 @@
-export type FetchLike = typeof fetch
