--- yarn/raijin/src/runtime/installer.interfaces.ts
@@ -1,6 +1,0 @@
-import type { FetchLike } from './download.interfaces.js'
-
-export interface InstallRaijinRuntimeOptions {
-  cwd: string
-  fetchImpl: FetchLike
-}
